+++ CHANGELOG.md
@@ -90,10 +90,17 @@ Note the following more specific changes to these related classes:
   the promises returned by the service.
 * Added alpha version of `DashContainer` for building dynamic, draggable dashboard-style layouts.
   Please note: the API for this component is subject to change - use at your own risk!
+* `Select` now allows the use of objects as values.
 * Added a new `xhEnableImpersonation` config to enable or disable the ability of Hoist Admins to
   impersonate other users. Note that this defaults to `false`. Apps will need to set this config to
   continue using impersonation. (Note that an update to hoist-core 6.4+ is required for this config
   to be enforced on the server.)
+* `FormField` now supports a `requiredIndicator` to customize how required fields are displayed.
+* Application build tags are now included in version update checks, primarily to prompt dev/QA users
+  to refresh when running SNAPSHOT versions. (Note that an update to hoist-core 6.4+ is required for
+  the server to emit build tag for comparison.)
+* `CodeInput` component added to provide general `HoistInput` support around the CodeMirror code
+  editor. The pre-existing `JsonInput` has been converted to a wrapper around this class.
 * AppBar now support an `appMenuButtonPosition` option to place the App Menu on the left or the right. 
 
 ### 💥 Breaking Changes
@@ -109,6 +116,12 @@ Note the following more specific changes to these related classes:
 * `TabRenderMode` and `TabRefreshMode` have been renamed to `RenderMode` and `RefreshMode` and moved
   to the `core` package. These enumerations are now used in the APIs for `Panel`, `TabContainer`,
   and `DashContainer`.
+* `DockViewModel` now requires a function, or a HoistComponent as its `content` param. It has always
+  been documented this way, but a bug in the original implementation had it accepting an actual
+  element rather than a function. As now implemented, the form of the `content` param is consistent
+  across `TabModel`, `DockViewModel`, and `DashViewSpec`.
+* `JsonInput.showActionButtons` prop replaced with more specific `showFormatButton` and
+  `showFullscreenButton` props.
 
 ### 🐞 Bug Fixes